requirements.txt
tensorflow
keras
numpy
tqdm
gradio
Pillow
nltk
kagglehub
matplotlib
